Move AppImage commands to AppImage namespace
