+++ Shelly-CLI/Commands/AppImage/AppImageSettings.cs
@@ -2,7 +2,7 @@ using System.ComponentModel;
 using Spectre.Console.Cli;
 using PackageManager.AppImage;
 
-namespace Shelly_CLI.Commands.Standard;
+namespace Shelly_CLI.Commands.AppImage;
 
 public class AppImageDefaultSettings : CommandSettings
 {
@@ -11,7 +11,7 @@ public class AppImageDefaultSettings : CommandSettings
     public bool Json { get; set; }
 }
 
-public class AppImageSettings : CommandSettings 
+public class AppImageSettings : CommandSettings
 {
     [CommandOption("-l | --location")]
     [Description("Location of the .AppImage to be installed")]
@@ -24,7 +24,7 @@ public class AppImageSettings : CommandSettings
     [CommandOption("-u|--update-url")]
     [Description("Set the release URL for update checking (e.g., https://github.com/owner/repo/releases)")]
     public string? UpdateUrl { get; set; } = "";
-    
+
     [CommandOption("-t|--type")]
     [Description("Set the update type (None, StaticUrl, GitHub, GitLab, Codeberg, Forgejo)")]
     public UpdateType UpdateType { get; set; } = UpdateType.None;
@@ -68,7 +68,7 @@ public class AppImageSyncMetaSettings : AppImageDefaultSettings
     [CommandArgument(0, "[QUERY]")]
     [Description("The search query for the AppImage")]
     public string? Query { get; set; } = "";
-    
+
     [CommandOption("-n|--no-confirm")]
     [Description("Proceed without asking for user confirmation")]
     public bool NoConfirm { get; set; }
